Fix: 회원 탈퇴 시 pet 삭제 전에 pet이 null인지 확인하는 코드 추가

+++ src/main/java/com/cocos/cocos/api/member/service/MemberService.java
@@ -263,6 +263,10 @@ public class MemberService {
     private void deleteAboutPet(final Long memberId) {
         final Pet pet = petRepository.findByMemberId(memberId);
 
+        if (pet == null) {
+            return;
+        }
+
         petSymptomRepository.deleteAllByPetId(pet.getId());
         petDiseaseRepository.deleteAllByPetId(pet.getId());
         petRepository.deleteById(pet.getId());
